Fix unmarshalling 64 bit map flags
We were storing the map flags in an `unsigned int` when unmarshalling
them which caused the MAP_MORE_ITEMS_GOOD and MAP_MORE_ITEMS_ARTEFACT
flags to be lost as they require more than 32 bits to store.

+++ crawl-ref/source/tags.cc
@@ -6626,7 +6626,7 @@ void marshallMapCell(writer &th, const map_cell &cell)
 void unmarshallMapCell(reader &th, map_cell& cell)
 {
     unsigned flags = unmarshallUnsigned(th);
-    unsigned cell_flags = 0;
+    uint64_t cell_flags = 0;
 
     cell.clear();
